Fix MAX_QUERY_BUF overflow

+++ pg_stat_monitor.h
@@ -93,7 +93,7 @@
 /* the assumption of query max nested level */
 #define DEFAULT_MAX_NESTED_LEVEL	10
 
-#define MAX_QUERY_BUF						(pgsm_query_shared_buffer * 1024 * 1024)
+#define MAX_QUERY_BUF						((long long)pgsm_query_shared_buffer * 1024 * 1024)
 #define MAX_BUCKETS_MEM 					((long long)pgsm_max * 1024 * 1024)
 #define BUCKETS_MEM_OVERFLOW() 				((hash_get_num_entries(pgsm_hash) * sizeof(pgsmEntry)) >= MAX_BUCKETS_MEM)
 #define MAX_BUCKET_ENTRIES 					(MAX_BUCKETS_MEM / sizeof(pgsmEntry))
